+++ bolatbek.nazgul/T2/ScopeGuard.cpp
@@ -1,4 +1,4 @@
-#include "ScopeGuard.hpp"
+#include "ScopeGuard.h"
 
 namespace bolatbek
 {
