--- bolatbek.nazgul/T2/ScopeGuard.hpp
@@ -1,21 +1,0 @@
-#ifndef SCOPEGUARD_HPP
-#define SCOPEGUARD_HPP
-
-#include <iomanip>
-
-namespace bolatbek
-{
-  class iofmtguard
-  {
-  public:
-    explicit iofmtguard(std::basic_ios< char > &s) noexcept;
-    ~iofmtguard();
-  private:
-    std::basic_ios< char > &s_;
-    char fill_;
-    std::streamsize precision_;
-    std::basic_ios< char >::fmtflags fmt_;
-  };
-}
-
-#endif
